Enforce nice and readable commit messages
Files changed:
* .gitmessage
* commit-lint.yml
* PULL_REQUEST_TEMPLATE.md
* R__refresh_user_stats_view.sql
* README.md
* setup-git-hooks.sh

Purpose of the change: (minimum 50 characters)
The whole sole purpose of this commit was to enforce some logical and readable commit messages. That might seem like a big overhead at this stage, but it will be extra helpful where there are lots of commits and lots of changes.

How does it affect the application: (minimum 50 characters)
Well, it doesn't affect the application per se, but it affects the developers and their whole experience while working on this project.

Additional notes:
Removed the version from the R migration(there is no need of versioning, it's repeatable). Also removed the .ci folder and the validate_undos.sh script that is pointless at this point.