Automatically populate files changed list
Files changed:
* .git-hooks/commit-msg
* .git-hooks/prepare-commit-msg
* README.md
* scripts/setup-git-hooks.sh

Purpose of the change: (minimum 50 characters)

The purpose of the change is to remove the confusion among the
developers. It automagically adds the list of changed files to the list
of files changed so on the developer is to just write meaningful
message.

How does it affect the application: (minimum 50 characters)

It does not affect the application, but it does affect the developers
working on it. It helps them with the actual development process and
everything that it involves.