Test the PR functionality added earlier
Files changed:
* README.md

Purpose of the change: (minimum 50 characters)

To test the whole behavior when trying to submit a PR to the develop
branch. This should do the trick, but let us try to be sure of the whole
process.

How does it affect the application: (minimum 50 characters)

It does not affect the application. It does not affect the developers
either. It just exists so I can try the PR functionality.