Make the PR logic not so strict
Files changed:
* .github/PULL_REQUEST_TEMPLATE.md

Purpose of the change: (minimum 50 characters)

To actually allow the developers to push code, not to get their code
rejected automatically for no paricular reason. Will test it soon.

How does it affect the application: (minimum 50 characters)

It affect only the developers and the development process. We will se
how it goes, but it should be better than earlier. Hopefully.