Remove the need to check old commits
Files changed:
* .github/workflows/commit-lint.yml

Purpose of the change: (minimum 50 characters)

To make the PR process even easier and more intuitive. Because the
current one is clunky and not doing the job it is supposed to do.

How does it affect the application: (minimum 50 characters)

It affect the developers by providing them with more flexibility when
creating commits and PRs. Will see how that goes.