Test the updated PR mechanism
Files changed:
* README.md

Purpose of the change: (minimum 50 characters)

To test the updated PR mechanism that now checks only the new commits
and does not care about the old ones. Here goes the try.

How does it affect the application: (minimum 50 characters)

If it works as intended it will then affect the developers very
positively. If not, i am afraid that it will get tricky...

Additional notes: