Change the lint again
Files changed:
* .github/workflows/commit-lint.yml

Purpose of the change: (minimum 50 characters)

To ignore the helper texts in order to make the PRs actually mergeable
so the developers can focus on their work, not on the stoopid commit
messages and PRs creation.

How does it affect the application: (minimum 50 characters)

It is supposed to benefit the developers. But i guess we will see soon
enough. The moment of truth is now.

Additional notes: