Change number who knows
Files changed:
* .github/workflows/commit-lint.yml

Purpose of the change: (minimum 50 characters)

The purpose is to skip the commits that include merges or something.
Just to validate the actual viable commits and their format.

How does it affect the application: (minimum 50 characters)

Maybe now actually allow the guys and girls to do their job. I guess we
will see how it goes.

Additional notes: