Fix R__refresh_user_stats_view.sql migration
Files changed:
* README.md
* src/main/resources/db/migration/repeatable/R__refresh_user_stats_view.sql

Purpose of the change: (minimum 50 characters)

The application failed to start because flyway was having trouble
altering the view(that was clearly my fault) so i fixed the migration.
Now works like a charm.

How does it affect the application: (minimum 50 characters)

Previously the application failed to start. Now it is alright and well.
Try ./gradlew devUp. #HappyCoding

Additional notes: