Optimize build.gradle.kts
Files changed:
* build.gradle.kts

Purpose of the change: (minimum 50 characters)

Optimize the custom tasks a bit. Add status and different prints. Add -d
to containers creation so you actually know when the service/containers
are up.

How does it affect the application: (minimum 50 characters)

The application now runs in the background(the containers actually) so
it is same as deployed on a server. Plus the feedback to the developer
that the task is actually successfull.

Additional notes: