Transfer the commit check to GitHub
On branch develop
Your branch is up to date with 'origin/develop'.

Changes to be committed:
  (use "git restore --staged <file>..." to unstage)
        deleted:    .git-hooks/commit-msg
        deleted:    .git-hooks/prepare-commit-msg
        deleted:    scripts/setup-git-hooks.bat
        deleted:    scripts/setup-git-hooks.sh

Changes not staged for commit:
  (use "git add <file>..." to update what will be committed)
  (use "git restore <file>..." to discard changes in working directory)
        modified:   .gitmessage

This change is supposed to allow the Windows developers to finally do
what they are supposed to do so. If there is a problem, please contact
me.

It mostly affects the development cycle of the application and makes the
lives of the developers a bit harder. Other than that, no probllama.