Change up stuff for Windows devs
Files changed: (optional)

 commit-lint.yml
 commit-message-push-lint.yml
 .gitmessage

Purpose of the change: (minimum 50 characters)
To enable the fellow Windows machines developers to actually contribute
to this dummy project. Hopefully i figured it out... It was about time

How does it affect the application: (minimum 50 characters)
It does not affect the application per se, but it does affect the
developers and the whole development process. Good luck and happy
coding!

Nope.!.