Update README.md
Files changed: (optional)

README.md

Purpose of the change: (minimum 50 characters)
To actually make the README.md more readable. And easier for new folks
to get on-board on the project.

How does it affect the application: (minimum 50 characters)
It does not affect the application whatsoever. It just provides insight
into what is going on in this projecto.