Add new migrations, learn JSONB
Files changed: (optional)
V5__create_users_metadata_table.sql
V6__create_vehicles_table.sql

V2__insert_initial_data.sql

Purpose of the change: (minimum 50 characters)
It is the next logical step to learn how to work with JSONB objects in
the database. That is the purpose of this commit.

How does it affect the application: (minimum 50 characters)
Huge impact on long run. No overhead with 5-10 entries per table. But if
it gets up to hundreds and thousands, indexes will be needed.